keyserver: (1) introduce epoche mechanism; (2) fix for keylist generation
(1) jmp_next_session, jmp_next_epoche, prepare_new_epoche
This functions allows asynchronous handling of "jumping to the next epoche"
and preparing the data for the NEXT epoche. This idea reflects the inticacies
of network communication, where on the one hand everybody operates
with the same data, on the other hand every node has to think of his future
communication and therefore assure that the data for the next epoche is
present in time.

(2) Fix keylist generation
keylist has to be cleaned when you keys are generated.