Make TCPFragmenter work correctly with FIN
Since TCPFragmenter copies all of a packet's headers to the fragmented
packets, each fragment inherits the original packet's flags.  This includes
the FIN flag, so if FIN were set each fragment will look like a FIN; this
prematurely terminates the flow.

This change modifies the behaviour to only allow a FIN on the last fragment.

Reviewed-By: cliff@meraki.com

git-svn-id: svn+ssh://svn.meraki.net/var/local/svnbase/meraki/trunk/router@102018 291e08f9-2c14-0410-acb0-c89b6de265c5