linuxmodule/todevice: fix logic for calling ndo_start_xmit
It seems much cleaner to me to directly call ndo_start_xmit rather than
dev_queue_xmit, as it avoids an extra layer of queueing and improves
backpressure behavior.  This should fix the NETDEV WATCHDOG issue that
Joonwoo fixed in d65bfe40, but in a different way.

In general, the logic in todevice.cc should be the same as
linux/net/core/pktgen.c.  This brings the code up to match that.