sync: use CLICK_ALWAYS_INLINE on SpinlockIRQ
According to LDD3, spin_lock_irqsave and spin_unlock_irqrestore can break on
some architectures if they are called from different functions.  Also, GCC
ignores inline if it is not optimizing.  This means that SpinlockIRQ is
broken

	1. in the linux module
	2. on those architectures
	3. if the compiler is not optimizing.

This change adds CLICK_ALWAYS_INLINE to SpinlockIRQ::{acquire,release} so
that these methods are inlined even if no optimization level is set.

Signed-off-by: Derrick Pallas <pallas@meraki.com>