sync/ReadWriteLock: account for actual cache line size
This code wants every Spinlock in the ReadWriteLock to occupy its own cache
line.  The way it is currently done is buggy.  First, L1 cache lines are not
always 32 bytes.  Second, if a Spinlock were larger than a cache line, the
code would not compile.

This change fixes the first issue by using L1_CACHE_BYTES and the second
with an anoymous union.

Signed-off-by: Derrick Pallas <pallas@meraki.com>