Be a little more cautious about click_max_cpu_ids
If there isn't support for thread-local storage (in userlevel), we rely on
sched_getcpu() in click_current_cpu_id() which could be upper than
click_nthreads (number of threads that click uses). As for the BSD mode we'll
use CLICK_CPU_MAX as an upper bound for the current cpu id which will be 256
when MT is enabled and 1 if not.